requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
db-dtypes==1.5.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pkg_resources==0.0.0
proto-plus==1.27.0
protobuf==6.33.4
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
